requirements.txt
flask==3.1.0
python-dotenv==1.0.1
langchain-groq==0.3.2
langgraph==0.3.21
duckduckgo-search==7.5.1
